Revert split of AbstractBlueprintIterator (2)
This reverts commit f1da773b82f1aadf3dd6d251127bf79ceeedcc02.
